Generate simple tables-of-contents upon request
To get a simple, two-level table of contents (TOC), include this line
right after the final header line:

    <div class="table-of-contents"/>

(This should be a NAV element, but Tidy mistakenly treats NAVs as inline
 elements and so wraps them in P elements, which is not helpful.)

To omit a section and all of its subsections from the TOC, annotate the
section’s header with `{toc=omit}`.  To omit all of the subsections of a
section but not the section itself, annotate the section’s header with
`{toc=omit-children}`.  For example:

    ## Style Guidelines for Text Blocks {toc=omit-children}